% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasm.R
\name{instantiate}
\alias{instantiate}
\title{Instantiate a WebAssembly module}
\usage{
instantiate(path)
}
\arguments{
\item{path}{the path to the binary wasm file}
}
\description{
Instantiate a WebAssembly module
}
