% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory-view.R
\name{new_memory_view}
\alias{new_memory_view}
\title{Create a new memory view}
\usage{
new_memory_view(wasm_instance_module, instance, offset)
}
\arguments{
\item{wasm_instance_module}{a pointer to the rcpp wasm module}

\item{instance}{an object of the Instance class}

\item{offset}{a positive integer or 0}
}
\description{
Create a new memory view
}
