% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-function.R
\name{typed_function}
\alias{typed_function}
\title{Use this to define a typed function}
\usage{
typed_function(fun, param_types, return_type = character())
}
\arguments{
\item{fun}{the function}

\item{param_types}{a character vector of wasm types with the length of the
numbers of parameters}

\item{return_type}{a character vector of length 0 or 1}
}
\description{
Use this to define a typed function
}
