% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance-class.R
\name{instantiate}
\alias{instantiate}
\title{Instantiate a WebAssembly module}
\usage{
instantiate(path, imports = list())
}
\arguments{
\item{path}{the path to the binary wasm file}

\item{imports}{a named list of import function}
}
\description{
Instantiate a WebAssembly module
}
